Change GxDynamicCall to GxClasses
