--- dotnet/src/dotnetcommon/GXDynamicCall/GxDynCallProperties.cs
@@ -1,28 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-
-namespace Genexus.DynamicCall
-{
-	public class GxDynCallProperties
-	{
-		public string ExternalName {
-			get;
-			set;
-		}
-		public string AssemblyName
-		{
-			get;
-			set;
-		}
-		public string NameSpace
-		{
-			get;
-			set;
-		}
-
-	}
-}
